+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
+### Changed
+
+- The argument to `:StopTimer` is now optional when there is only one active timer.
+
 ## [v0.1.0](https://github.com/epwalsh/pomo.nvim/releases/tag/v0.1.0) - 2023-11-29
 
 ### Added